expose `ThreadPool::global()` function
This allows you to get a `ThreadPool` representing the global
thread-pool. We give back a `&'static Arc<ThreadPool>`, quite
intentionally:

- This allows you to copy the global thread-pool around without
  any ref-counting, using `&Arc<ThreadPool>` or `&ThreadPool`;
- You can `clone()` this to get a `Arc<ThreadPool>` which is perhaps
  convenient, particularly if you want to intermingle with thread-pools
  you may have created.

In particular this means things like
`ThreadPool::global().spawn(...)` would be very fast (no
unnecessary ref-counting).

NAMING QUESTION: `global()` or `default()`? I chose the former because I
didn't this to be confused with the `Default` trait, and implementing
`Default` for `ThreadPool` seems like not obviously a good idea (but
also not obviously bad).